// packages/kms-azure-rest-client/src/js-client/models/CreateEcKeyRequest.ts
/**
 * Azure KeyVault REST API
 * REST API for interacting with Azure Key Vault
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateEcKeyRequest {
    'keyName'?: string;
    'curveName'?: string;
    'operations'?: Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "keyName",
            "baseName": "keyName",
            "type": "string",
            "format": ""
        },
        {
            "name": "curveName",
            "baseName": "curveName",
            "type": "string",
            "format": ""
        },
        {
            "name": "operations",
            "baseName": "operations",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateEcKeyRequest.attributeTypeMap;
    }

    public constructor() {
    }
}